Added Peter's PR comments to codebase as TODO: comments. Moved epsilon_bipartition_tree() from tree.py to tree_proposals.py Moved recursive_tree_part() and recursive_seed_part() to a new module partition/initial_partition_generators.py All tests pass...
