+++ gerrychain/metrics/compactness.py
@@ -1,6 +1,8 @@
 import math
 from typing import Dict
 
+# frm: TODO: Documentation: Add import of Partition just to be clear...
+
 
 def compute_polsby_popper(area: float, perimeter: float) -> float:
     """
