+++ gerrychain/partition/__init__.py
@@ -1,4 +1,5 @@
 from .geographic import GeographicPartition
+from .initial_partition_generators import recursive_seed_part, recursive_tree_part
 from .partition import Partition
 
-__all__ = ["Partition", "GeographicPartition"]
+__all__ = ["Partition", "GeographicPartition", "recursive_tree_part", "recursive_seed_part"]
